LF-15098-10 ptp: netc: Make only ALARM1 available on i.MX95 platform
Currently, the NETC Timer needs to be configured in FS mode so that the
PPS and PEROUT pulses can be generated at the expected time, for example,
the PPS signal is always generated when the PHC reaches 1-rounded second.
The PEROUT pulse needs to be generated at the specified time.

These expected times are achieved by configuring the ALARM registers, but
for i.MX95, only ALARM1 supports FS mode. So only ALARM1 is available on
i.MX95 platform. And i.MX94 platform does not have such limitation.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>