LF-15025-6 net: pcs: xpcs: clear the MAC_AUTO_SW bit when configuring 10g-baser mode
When only a 10G link is available, the MAC_AUTO_SW bit will not be set.
However, if the PHY supports both SGMII and 10G-baser simultaneously, after
the link being configured as 1G SGMII, the MAC_AUTO_SW bit will be set.
When switching back to a 10G-baser link, due to the MAC_AUTO_SW bit not
being cleared, the link will fail.

So clear the MAC_AUTO_SW bit when config the link to 10g-baser.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>