LF-15098-11 ptp: netc: re-enable FIPER when adjusting the current time
If FIPER is not disabled when adjusting the time, FIPER will continue to
output pulse, which will cause deviations. The pulse will not be aligned
with the expected time until the reset ALARM time takes effect. For
example, using phc_ctl to set the current time of PHC will cause a large
deviation in the PPS pulse.

$ phc_ctl /dev/ptp0 set 154.500500000
event index 0 at 155.284095092
event index 0 at 157.000000018
event index 0 at 158.000000018

So it is better to disable FIPER before adjusting the current time and
then re-enable FIPER to avoid the unexpected PPS signal. The result is
shown as follows.

$ phc_ctl /dev/ptp0 set 154.500500000
event index 0 at 157.000000017
event index 0 at 158.000000017
event index 0 at 159.000000017

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>