LF-15098-12 ptp: netc: optimize the alarm time
In the current implementation, the alarm time is greater than or equal to
100ms than the current time. In fact, this is not necessary, because all
processing is always protected by spinlock_irqsave(), and the processing
time is very short, at the microsecond level. Therefore, there is no need
to set the alarm time larger, 1ms larger than the current time is enough.
In fact, 1ms is also larger, but we don't need to be picky and set a
smaller value.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>