LF-15269 net: enetc: specify gate operation type for Qbv gate entry
IEEE 802.1Q-2022 has defined the option of using time gate scheduling
in combination with HOLD/RELEASE and preemption, to protect the express
traffic class window from interference from preemptible traffic, while
at the same time minimizing the impact of the protected window on the
amount of bandwidth that is available to preemptible traffic. To support
this option, the standard has defined the "Set-And-Hold-MAC" and
"Set-And-Release-MAC" gate operations in a gate control list, to specify
the start and the end of an express traffic class protected window.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>