LF-15336 net: stmmac: synchronize with the tx transmission when stopping the queues
If stmmac_suspend() is called asynchronously during data transmission and
the queue has been checked and is not stopped or freezed, it will cause
the stmmac clocks to be turned off before waiting for the tx transmission
to complete.
To fix this, netif_tx_lock_bh() is called before netif_device_detach() to
synchronize with the tx race.

At the same time, disable NAPI first to ensure that NAPI handlers are not
running. Similar sequence adjustments are also made during the resume
process.

Additionally, stmmac_disable_all_queues() and netif_device_detach() are
removed from mutex_unlock(&priv->lock) scope in both suspend and resume
funcitons, because they call the common interfaces and are not protected
by this lock at other call point either.

Fixes: 874bd42d24c2 ("stmmac: convert to dev_pm_ops.")
Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Acked-by: Wei Fang <wei.fang@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>