LF-15623-10 net: dsa: netc: restore tc-flower tables when switch resumes
If the NETC is powered off in suspend mode, all the configurations of
the tc-flower tables will be lost. Therefore, when the system resumes,
we need to restore the configurations to ensure that tc-flower can
still resume working after the system resumes.

In addition, since multiple rules may share the same gate/police/key
table entries, so the restored flag is added to indicate the entry has
been restored, no need to set it up again.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Acked-by: Clark Wang <xiaoning.wang@nxp.com>