LF-15671-1 net: dsa: netc: correct entry index of ETT and ECT
For i.MX943 NETC switch, every four consecutive entry indexes form an
egress treatment group, and the same is true for egress counter table
(ECT). That is, for each VLAN, each port has a corresponding ETT entry,
and each ETT entry corresponds to an ECT entry.

However, there is an error when adding a new ETT entry, the ec_eid of
ett_cfge is mistakely set to ett_eid rather than ect_eid. In addition,
when deleting an ETT group, the deleted indexes are not continuous
because ett_eid is incorrectly set to "ett_eid += i;".

Another issue is that ett_eid_bitmap and ect_eid_bitmap both represent
the group's bitmap. So when clearing the bitmap, the group index should
be used instead of the entry index.

Fixes: 5b5e66e99332 ("LF-14260-12 net: dsa: netc: add basic switch functions support")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>