LF-15671-4 net: dsa: netc: correct phy_mask for the internal MDIO bus
The internal MDIO bus is created to control the PCS, which is not a PHY
device, so the phy_mask should be 0xFFFFFFFF. Otherwise, 32 PHY devices
will be created by mdiobus_register(), see below.

mdiobus_register() --> mdiobus_scan_bus_c45()
			--> mdiobus_scan()
				--> phy_device_register()

Therefore, this patch is used to correct the phy_mask for the internal
MDIO bus.

Fixes: 2349f9607f2a ("LF-14260-10 net: dsa: netc: introduce NXP NETC Switch driver for i.MX94")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>