LF-15671-7 net: dsa: netc: save police_tbl to rule->police_tbl
To make the same police action can be shared by multiple flower rules,
the police_tbl pointer should be saved to the rule to ensure that the
driver does not repeatly create the same RPT entry or mistakely delete
the RPT entry that is still in use. However, we forgot to save the
police_tbl in netc_setup_trap_redirect(), which will undoubtedly cause
problems. So this patch is used to fix this problem.

Fixes: 81c86ba05455 ("LF-14819-2 net: enetc: make police action can be shared by multiple rules")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>