LF-15671-10 net: dsa: netc: remove netc_port_set_mac_address()
For NETC switch ports, the PMAR0/1 registers are not used to filter MAC
address, the packets received by the switch port are forwarded to CPU
port by matching FDB. These two registers are used when modifying the
packet, such as changing the SMAC of the packet to the MAC address in
PMAR0/1. However, the current driver does not support Frame Modification
Table, so setting PMAR0/1 is useless now, so it is safe to remove
netc_port_set_mac_address(). And we can add them back when we really
need them.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>