LF-15703 net: enetc: add a path to get the MAC address from netdev
When the system resumes, if U-Boot has not updated the MAC address in
the DTB and the fuse is empty, it may lead to the MAC being reassigned
a random address after resume.

To maintain the consistency of the MAC address before and after
suspend/resume, add this path to obtain the MAC address from the netdev.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>