LF-15779-1 net: enetc: correct the entry index when setting MAFT entry
The MAC address filter table (MAFT) only has 4 entries, and the number
of MAC addresses of MAC filter is more than 4 when both PF and VFs are
enabled. So at this time we should use MAC hash filter instead of MAFT.
However, there is a problem with the current implementation. When the
number of current MAC addresses exceeds 4, it incorrectly determines
that the current MAFT is still available, resulting in access to a
non-existent entry. Therefore, enetc_mac_list_is_available() is added
to determine whether the current MAFT is available.

Fixes: 93eab37d2501 ("LF-11845-1 net: enetc: convert mailbox messages to new formats")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>