LF-15779-3 net: enetc: use MAFT to filter unicast only
Currently, the MAC address filter table (MAFT) only has 4 entries, and
the table is shared by PF and VFs. However, the MAC filter is configured
with multiple multicast addresses by default. So the number of multicast
addresses will exceed 4 if both PF and VFs are enabled. Therefore, MAFT
is not suitable for multicast filtering. So MAFT is better to be used to
filter unicast only. If the number of unicast addresses exceeds the table
capacity, the MAC hash filter will be used.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>