imx8ulp-clock: add TPM6 and TPM7 clock selectors
To select the clock source for TPM6 and TPM7 of the i.MX 8ULP, not only
the respective PCC registers but also the TPM6_7CLK register of the
AD_CGC module must be set.

Since the default value of the TPM6_7CLK register refers to the clock
generated by the Audio PLL, the TPM6_7CLK register must be set
explicitly in order to use the clock selected in the PCS field of the
PCC register as the source.

However, unlike the SAI5_4_CLK register, which has a similar function,
the TPM6_7CLK register was not defined as a constant in imx8ulp-clock.h
or clk-imx8ulp.c to allow such a specification.

Add the definition of the TPM6_7CLK register to these definition files
in order to use the TPM7 as a PWM with the source clock selectable in
the PCS field of the PCC register.

The above issue is also discussed in the NXP forum below.
https://community.nxp.com/t5/i-MX-Processors/File-drivers-clk-imx-clk-imx8ulp-c-is-missing-some-registers/m-p/2068008?profile.language=ja

Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>
(cherry picked from commit 3907764b748b7b33cab36e2c229c406f3b3e40a0 @ https://github.com/atmark-techno/linux-5.10-at)
Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>