LF-15909 net: enetc: fix the behavior difference between iMX95 A1 and iMX95 B0/iMX943
On iMX95 B0 and iMX943, different behaviors were observed compared to
iMX95 A1.
If the PIR value of the Tx BD ring is updated when the link is down
by only disabling PM_CMD_CFG_TX_EN, no issue occurs on iMX95 A1
and the CIR will be still updated.
However, on iMX95 B0 and iMX943, this situation causes an error in
the CIR of the Tx BD ring, resulting in the inability to recover this BD.

This difference can be resolved by adding writing the POR register
when the link is down.

Since it is noted on RM that when disabling the Tx BD ring, it is
necessary to ensure that PIR = CIR, drop the packets when the link
is down to avoid further potential issues with the BD ring.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>