LF-15907-3 arm64: dts: imx95: rename the dts file name to support tja1104
TJA1103-SDBR will be replaced by TJA1104-SDBR, so rename the file name from
tja1103 to tja1104 for imx95.
TJA1104 supports all the functions of TJA1103, and in addition, it also
supports the MacSec function.

Remove the tja1120, because it is not supported.

Update the rework description of imx95-19x19-evk-tja1104-rmii.dts.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>