phy: lynx-10g: fix protocol change towards 10G on LS1046A lane C
It was reported that on LS1046A, transitioning a lane towards 10G
Ethernet results in broken traffic on lane C, but is functional on
lane D.

Analyzing the LS1046A SerDes register documentation, it became apparent
that there exists an irregularity where for lane C, the PCCRB[XFIA_CFG]
field needs to be written with 0b010, but for lane D, the PCCRB[XFIB_CFG]
field needs to be written with 0b001.

The driver is currently completely unaware of such complexities, and
always writes 0b001 to PCCRB[XFI*_CFG], which is fine for all other SoCs
and lanes.

Introduce a "pccr_override" function pointer, which is implemented only
for LS1046A SerDes 1, in order to return 0b010 for 10G Ethernet. For all
other configurations, return -EOPNOTSUPP, which is the same as not
implementing the "pccr_override" method at all, which in turn is the
same as letting the default code path write PCCRB_XFIa_CFG. This is fine,
because no other SerDes seems to need to override the default value.

Fixes: e0e242341165 ("phy: lynx-10g: new driver")
Reported-by: Gopi Krishna M <gkrishna@lantronix.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Acked-by: Wei Fang <wei.fang@nxp.com>