LF-15623-7 net: dsa: netc: improve suspend and resume feature
The initial patch is only to solve some crashes caused by the switch
driver. So that the switch driver will not block or break the system
suspend and resume process. Therefore, the initial patch does not
actually ensure that the switch will work properly after resume.

When the system enters the suspend mode, NETC has two states, one is
powered off, the other is still powered on. Currently we only know
that NETC will not be powered off in suspend mode when WOL is enabled.
For other cases, there is no helper function for us to query whether
NETC is powered off in suspend mode. So we assume that NETC is powered
off in suspend mode, and all hardware configurations will be lost in
suspend mode. Therefore, the driver needs to restore the configurations
when the switch resumes.

For the case that NETC is powered off in suspend mode, currently, we
have completed the recovery of registers, VLAN filter table, egress
treatment table (ETT) and egress counter table (ECT), FDB table, and
PTP related IPFT entries. However, there are still some tables that
have not been restored, such as time gate scheduling table (TGST) and
tc-flower related tables, such as ingress stream sdentification table
(ISIT), ingress stream table (IST), stream gate instance table (SGIT),
etc. The recovery of these tables will be supported in subsequent
patches.

For the case that NETC is not powered off in suspend mode (WOL is
enabled), it is relatively simple, because all hardware configurations
will not be lost, so we only need to disable the switch PCIe device when
the system suspends, disable the related clocks and work queue, and then
re-enable these functions when the system resumes.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>