LF-15623-5 net: enetc: initialize sw PIR/CIR by reading hw PIR/CIR
Software can only set PIR and CIR registers after a reset, like FLR or
warm reset, and these two registers can only be set to 0. But these two
registers reset value are 0, so actually software does not need to set
them.

In addition, there is a logic error in the current implementation,
next_to_clean indicates the software CIR, it corresponds to the hardware
CIR, and next_to_use indicates the software PIR, it corresponds to the
hardware PIR. But we use next_to_clean to set PIR and use next_to_use to
set CIR.

Therefore, this patch removes the setting for CIR and PIR and initializes
next_to_clean and next_to_use by reading CIR and PIR.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>