LF-16293 net: enetc: add VF suspend/resume supoort
Currently, the VF driver does not support suspend/resume functionality.
The PF driver disables SR-IOV when the system suspends and then enables
it when the system resumes. Therefore, the VF driver needs to unbind
and rebind the device during each suspend/resume cycle. As a result,
all VF configurations are lost and cannot be restored because the VF
driver is removed. Therefore, it is best to add suspend/resume support
to the VF driver, and let the VF driver restore the configuration when
the system resumes.

It is worth noting that the PF driver and VF driver currently do not
establish a supplier-consumer relationship. This means the PF driver
may suspend before the VF or resume after the VF, which can lead to
fatal issues. Therefore, a link is established between the PF and VF
drivers using device_link_add() to ensure that the PF driver suspends
after the VF and resumes before the VF.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>