LF-15227 drm/imx: dcnano-crtc: Fixup vpos in ->get_scanout_position() callback
If vpos happens to be the last line of active scanout region, enforce
vpos to be the first line of vblank region in >get_scanout_position()
callback.  This is needed because vblank core only supports calculating
vblank timestamp for the case where vblank IRQ is fired off at the end
of vblank region, while DCNANO fires it off at the end of active display
region.  This prevents vblank core from generating backward vblank
timestamps for consecutive frames.  This fixes weston assertion on
backward timestamp when comparing timestamps for consecutive frames.

Tested-by: Jared Hu <jared.hu@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>