LF-16031 net: enetc: fix crash issue if PF suspends with SR-IOV enabled
When ENETC PF suspends and its netdev is not running, the driver will
call enetc_free_msix() to free all previously allocated IRQ vectors.
However, if the PF has enabled SR-IOV, it will cause a call trace, the
log is as follows.

[   38.471290] remove_proc_entry: removing non-empty directory 'irq/257', leaking at least 'eth0-vfmsg'
[   38.480495] WARNING: CPU: 5 PID: 58 at /fs/proc/generic.c:725 remove_proc_entry+0x1a4/0x1b4
[   38.626124] Call trace:
[   38.628577]  remove_proc_entry+0x1a4/0x1b4
[   38.632686]  unregister_irq_proc+0xe0/0x118
[   38.636889]  free_desc+0x50/0xe0
[   38.640137]  irq_free_descs+0x68/0x98
[   38.643805]  irq_domain_free_irqs+0x16c/0x1c4
[   38.648180]  msi_domain_free_locked.part.0+0x178/0x1c8
[   38.653321]  msi_domain_free_irqs_all_locked+0x64/0x90
[   38.658469]  pci_msi_teardown_msi_irqs+0x48/0x5c
[   38.663097]  pci_free_msi_irqs+0x18/0x40
[   38.667032]  pci_free_irq_vectors+0x64/0xa0
[   38.671235]  enetc_free_msix+0xe8/0x100
[   38.675083]  enetc4_pf_power_down+0x50/0x7c
[   38.679278]  enetc4_pf_suspend+0x124/0x164
[   38.683386]  pci_pm_suspend+0x7c/0x260

And then there is a crash when the system resumes, the log is as follows.

[   39.813596] Kernel panic - not syncing: Asynchronous SError Interrupt
[   39.813657] Call trace:
[   39.813664]  dump_backtrace.part.0+0xd4/0xe0
[   39.813688]  show_stack+0x18/0x30
[   39.813707]  dump_stack_lvl+0x60/0x80
[   39.813727]  dump_stack+0x18/0x24
[   39.813744]  panic+0x168/0x360
[   39.813760]  add_taint+0x0/0xbc
[   39.813779]  arm64_serror_panic+0x64/0x70
[   39.813793]  arm64_is_fatal_ras_serror+0x3c/0xb0
[   39.813807]  do_serror+0x5c/0x70
[   39.813819]  el1h_64_error_handler+0x30/0x54
[   39.813838]  el1h_64_error+0x64/0x68
[   39.813852]  __pci_write_msi_msg+0x1d4/0x204
[   39.813873]  __pci_restore_msix_state+0x16c/0x188
[   39.813887]  pci_restore_msi_state+0x20/0x34
[   39.813907]  pci_restore_state.part.0+0x184/0x344
[   39.813929]  pci_restore_state+0x1c/0x34
[   39.813948]  pci_pm_resume_noirq+0x114/0x1c0

The first call trace is that the 'eth0-vfmsg' IRQ is not freed, so
remove_proc_entry() will report a warning if remove a non-empty
directory.

The second call trace is that the current VF driver does not support
suspend/resume, and the SR-IOV is not disabled when the PF suspends, so
the IRQ vectors of VF is not freed when system suspends, and the msi
driver will try to restore them when system resumes, but the bus master
enable bit of VF may be cleared, so any memory or I/O requests are not
allowed which will lead to an Asynchronous SError Interrupt.

Since the VF driver does not support suspend/resume at present, a quick
workaround for this issue is to disable SR-IOV when PF suspends (if its
netdev is not enabled). And SR-IOV is re-enabled when PF resumes. This
solution is same as when PF's netdev is enabled but WOL is not.

Fixes: 58fd162d4015 ("LF-11880-2 net: enetc: Add power off suspend/resume and WOL supports")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>