LF-16086 net: dsa: netc: fix incorrect port speed issue
The purpose of adding the speed check is to avoid reading and writing
NETC_PCR if the speed is not changed. However, it does not take into
account the fact that after suspend/resume, the switch will be reset.
Therefore, the port->speed and NETC_PCR[PCR_PSPEED] may be different.
If the speed parameter passed to netc_port_set_speed() is equal to
port->speed, it will cause NETC_PCR[PCR_PSPEED] to not be updated. To
solve this issue, the check is removed.

Fixes: 5fa1b6d2cd93 ("LF-15623-7 net: dsa: netc: improve suspend and resume feature")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>