LF-16332-7 net: enetc: perform WOL configuration only on ENETC with WOL enabled
In the current implementation, as long as one ENETC is configured with
WOL enabled, all other ENETCs will active PME and enable magic packet
detection and so on. This is not correct, and actually the WOL feature
of other ENETCs will not take effect because no corresponding IPFT entry
is configured. So correct this behavior for WOL-disabled ENETCs.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>