LF-16243 net: dsa: netc: allocate dedicates buffer pools for each port
In the current implementation, all the ports share all the buffer pools,
However, this may lead to performance degradation in some cases.

For example when the i.MX943 switch is configured in bridge mode, and
two Spirent TestCenter (STC) ports connects to the two SGMII ports of
the switch and then transmits traffic on these two STC port at a rate
of 2.5Gbps. Sometimes after a period of time, most of the packets in
one direction of the flow are discarded by the switch because of
congestion. This case is most likely to happen is that one of the PHYs
is a bit slower than the Spirent clock on the Tx side.  Eth spec allows
interfaces a +/-100ppm margin in terms of clocking, so this could happen.
This means that NETC switch constantly gets a little bit more BW than it
can put out at least for one of the flows, and this leads to buffers
accumulating slowly on the egress queue of the port that's a bit slower
to Tx. Because the rate difference is still relatively small, it would
take minutes for the queue to suck up all the buffers. Once the queue
did block all NETC switch buffers though, the switch is running on 0-1
free buffers, so we will see Rx discard (no buffers to put the Rx frames
in) on all ports (because the buffer pool is shared).

What is worse is that when this happens, the CPU cannot send packets
because the flow control between the internal ENETC and the switch CPU
port is always enabled. Therefore, when the switch is congested, the
internal ENETC will not send packets to the switch.

As per the IP team's recommendation, software needs to allocate dedicated
buffer pools to each port. Having a dedicated per-port buffer pool will
limit frame drops to flows where Tx cannot keep up.

In addition, based on the NETC block guides of other SoCs, the number of
ports, number of buffer pools, and the shared memory buffer sizes vary.
Therefore, the initialization of buffer pool table is better tailored to
the platforms. hence the bpt_init() is added to netc_switch_info so that
each platform could implement its own callback. Furthermore, the flow
control of each port is also implemented based on the buffer pool, so
the port_tx_pause_config() is also added to struct netc_switch_info.

Fixes: 7946e38e2c4a ("LF-14260-13 net: dsa: netc: add buffer pool support")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Acked-by: Clark Wang <xiaoning.wang@nxp.com>