LF-16283: arm64: dts: imx95-15x15-frdm: Change waveshare 7inch panel I2C assignment
The waveshare panel uses dupont line to connect its I2C interfrace
to i2c6 which is on the FRDM-IMX95 40-pin EXPI header.

At the begnning, the i2c4 is unavailable for dupont line connection.
Hardware requests to change the I2C connection from i2c6 to i2c4.
Since i2c6 is located on the 40-pin EXPI header, which is
commonly used by accessories such as the aud-hat.

To enable connection of additional accessories without conflict,
using i2c4 which is not on the 40-pin EXPI header and remains
available as new I2C interface.

Signed-off-by: Joseph Guo <qijian.guo@nxp.com>
Acked-by: Liu Ying <victor.liu@nxp.com>