LF-16264-10: media: ox05b1s: Make clock mandatory
The clock is now a required property in the sensor driver
bindings, to enable a complete and accurate hardware description,
so print an error instead of warning in probe.

Drop clock name, it is not needed in case of single clock.

Use dev_err_probe instead of dev_warn for missing clock.

Signed-off-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>