LF-16332-8 net: enetc: move rtnl_lock() before calling netif_running()
The netif_running() result can change state concurrently, it should be
better to acquire rtnl_lock() first to prevent that from happening.

In addition, enetc4_sriov_suspend() cannot be used in the rtnl_lock
context because it disables SR-IOV and causes the VF driver to acquire
the rtnl_lock, which will cause a deadlock. Therefore, move it before
acquiring rtnl_lock.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>