LF-16395: media: i2c: Move ox05b1s from "Video decoders" to VIDEO_CAMERA_SENSOR
The ox05b1s driver is no longer built in Linux 6.17 because the
"Video decoders" section in media/i2c/Kconfig is only visible when
CONFIG_MEDIA_HIDE_ANCILLARY_SUBDRV is unset. This option was not set in
linux-lts-nxp 6.12, but is now enabled in linux-nxp 6.17, revealing that
the Kconfig placement for ox05b1s was incorrect all along.

Move the ox05b1s driver to the VIDEO_CAMERA_SENSOR section, which is
its proper location.

Signed-off-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Tested-by: Antoine Bouyer <antoine.bouyer@nxp.com>